requirements.txt
fastapi
uvicorn
exchangelib
python-dotenv
